Add Inbox feature

+++ app/src/main/java/com/hmatalonga/greenhub/models/data/CpuStatus.java
@@ -27,8 +27,8 @@ public class CpuStatus extends RealmObject {
     public double cpuUsage;
 
     // Uptime in seconds
-    public double upTime;
+    public long upTime;
 
     // Experimental sleep time
-    public double sleepTime;
+    public long sleepTime;
 }
